build: export src/box/box.cc symbols
Exports functions visible to the public C API and used by
src/box/lua/schema.lua via FFI.

It is worth mentioning why box_export_syms() does not aggregate
all the other box_*_export_syms(). This is because there is a
broken dependency between main.cc and 'box' static lib.

'box' uses functions 'tarantool_uptime()' and 'load_cfg()', which
are defined in main.cc. When box_export_syms() tries to aggregate
symbols from its submodules, this break explodes. And its fix is
not as a trival as it seems, so it is postponed.

Part of #2971